Change `checkAuthorization` function
When checking for IDORs or similar vulnerabilities, it's important to consider that certain actions, such as deleting a post or voting on a post, may only be performed once. Sending the original request is likely to succeed, whereas unauthorized or modified requests may not work as the original has already been executed (for example: the post has been deleted).